refactor(panel-layout): extract HTML builders to layout/html.ts
The HTML-template portion of the 1,919-line panel-layout.ts was essentially a
self-contained rendering module — five methods producing template strings for
the macOS sidebar shell, the web-variant header shell, the map section, the
sidebar navigation, and the theme icon / sidebar update button.

Extracted all five into src/app/layout/html.ts as pure functions that take
the AppContext argument explicitly. panel-layout.ts keeps three thin
delegating wrappers for the ones still called from inside the class
(buildSidebarUpdateBtnHtml, buildDesktopLayout, buildWebLayout).

Net effect on panel-layout.ts: 1,919 → 1,702 lines (−217 lines / −11%).

Updated tests/panel-height-defaults.test.mjs to scan both panel-layout.ts
and the new layout/html.ts so its grid-vs-map layout contract assertions
still match.

Verification:
- typecheck:all clean
- test:sidecar 78/78
- test:api 36/36
- test:data 554/574 (same 20 pre-existing drift failures)
